requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.3
MarkupSafe==2.1.2
sqlalchemy~=2.0.35
WTForms~=3.1.2
mysql-connector-python
PyMySQL
mysqlclient